moved MetricsConnectionSettingsModule to Bigtable module since...
that's where it belongs, basically.
